+++ src/DependencyInjection/ServerDependencyScopes/App.razor
