+++ src/DependencyInjection/ServerDependencyScopes/MyTransientService.cs
@@ -1,21 +1,21 @@
 ﻿using System;
 
-namespace DependencyScopes
+namespace ServerDependencyScopes
 {
-	public interface IMySingletonService
+	public interface IMyTransientService
 	{
 		public TimeSpan DeltaCreationTime { get; }
 		public int InstanceNumber { get; }
 	}
 
-	public class MySingletonService : IMySingletonService
+	public class MyTransientService : IMyTransientService
 	{
 		public TimeSpan DeltaCreationTime { get; }
 		public int InstanceNumber { get; }
 
 		private static volatile int PreviousInstanceNumber;
 
-		public MySingletonService()
+		public MyTransientService()
 		{
 			DeltaCreationTime = DateTime.UtcNow - AppLifetime.StartTimeUtc;
 			InstanceNumber = System.Threading.Interlocked.Increment(ref PreviousInstanceNumber);
