+++ src/DependencyInjection/ServerDependencyScopes/Pages/Error.razor
