+++ src/DependencyInjection/ServerDependencyScopes/Pages/Index.razor
