+++ src/DependencyInjection/ServerDependencyScopes/Pages/_Host.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@namespace DependencyScopes.Pages
+@namespace ServerDependencyScopes.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
     Layout = null;
@@ -10,7 +10,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>DependencyScopes</title>
+    <title>ServerDependencyScopes</title>
     <base href="~/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
     <link href="css/site.css" rel="stylesheet" />
