+++ src/DependencyInjection/ServerDependencyScopes/Program.cs
@@ -9,7 +9,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace DependencyScopes
+namespace ServerDependencyScopes
 {
 	public class Program
 	{
