+++ src/DependencyInjection/ServerDependencyScopes/ServerDependencyScopes.csproj
