+++ src/DependencyInjection/ServerDependencyScopes/Shared/MainLayout.razor
