+++ src/DependencyInjection/ServerDependencyScopes/Shared/MyStandardComponent.razor
