+++ src/DependencyInjection/ServerDependencyScopes/Shared/NavMenu.razor
@@ -1,5 +1,5 @@
 ﻿<div class="top-row pl-4 navbar navbar-dark">
-	<a class="navbar-brand" href="">DependencyScopes</a>
+	<a class="navbar-brand" href="">ServerDependencyScopes</a>
 	<button class="navbar-toggler" @onclick="ToggleNavMenu">
 		<span class="navbar-toggler-icon"></span>
 	</button>
