+++ src/DependencyInjection/ServerDependencyScopes/Startup.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace DependencyScopes
+namespace ServerDependencyScopes
 {
 	public class Startup
 	{
