+++ src/DependencyInjection/ServerDependencyScopes/_Imports.razor
@@ -5,5 +5,5 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using DependencyScopes
-@using DependencyScopes.Shared
+@using ServerDependencyScopes
+@using ServerDependencyScopes.Shared
