+++ src/DependencyInjection/ServerDependencyScopes/appsettings.Development.json
