+++ src/DependencyInjection/ServerDependencyScopes/appsettings.json
