+++ src/DependencyInjection/ServerDependencyScopes/wwwroot/css/open-iconic/README.md
