+++ src/DependencyInjection/ServerDependencyScopes/wwwroot/css/open-iconic/font/fonts/open-iconic.ttf
