+++ src/DependencyInjection/ServerDependencyScopes/wwwroot/favicon.ico
