+++ src/DependencyInjection/WebAssemblyDependencyScopes/AppLifetime.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace WebAssemblyDependencyScopes
+{
+	public static class AppLifetime
+	{
+		public static DateTime StartTimeUtc { get; } = DateTime.UtcNow;
+		public static TimeSpan ElapsedTime => DateTime.UtcNow - StartTimeUtc;
+	}
+}
