+++ src/DependencyInjection/WebAssemblyDependencyScopes/MyTransientService.cs
@@ -1,0 +1,24 @@
+﻿using System;
+
+namespace WebAssemblyDependencyScopes
+{
+	public interface IMyTransientService
+	{
+		public TimeSpan DeltaCreationTime { get; }
+		public int InstanceNumber { get; }
+	}
+
+	public class MyTransientService : IMyTransientService
+	{
+		public TimeSpan DeltaCreationTime { get; }
+		public int InstanceNumber { get; }
+
+		private static volatile int PreviousInstanceNumber;
+
+		public MyTransientService()
+		{
+			DeltaCreationTime = DateTime.UtcNow - AppLifetime.StartTimeUtc;
+			InstanceNumber = System.Threading.Interlocked.Increment(ref PreviousInstanceNumber);
+		}
+	}
+}
