+++ src/DependencyInjection/WebAssemblyDependencyScopes/_Imports.razor
@@ -1,0 +1,9 @@
+﻿@using System.Net.Http
+@using System.Net.Http.Json
+@using Microsoft.AspNetCore.Components.Forms
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.AspNetCore.Components.WebAssembly.Http
+@using Microsoft.JSInterop
+@using WebAssemblyDependencyScopes
+@using WebAssemblyDependencyScopes.Shared
