+++ src/DependencyInjection/WebAssemblyDependencyScopes/wwwroot/sample-data/weather.json
@@ -1,0 +1,27 @@
+﻿[
+  {
+    "date": "2018-05-06",
+    "temperatureC": 1,
+    "summary": "Freezing"
+  },
+  {
+    "date": "2018-05-07",
+    "temperatureC": 14,
+    "summary": "Bracing"
+  },
+  {
+    "date": "2018-05-08",
+    "temperatureC": -13,
+    "summary": "Freezing"
+  },
+  {
+    "date": "2018-05-09",
+    "temperatureC": -16,
+    "summary": "Balmy"
+  },
+  {
+    "date": "2018-05-10",
+    "temperatureC": -2,
+    "summary": "Chilly"
+  }
+]
